Moved header files to ethercat_interface directory.
