+++ .vscode/c_cpp_properties.json
@@ -5,7 +5,7 @@
             "includePath": [
                 "${workspaceFolder}/**",
                 "/home/naci/ethercat/include/**",
-                "/home/naci/Projects/igh_master/include/"
+                "/home/naci/Projects/igh_master/include/**"
             ],
             "defines": [],
             "compilerPath": "/usr/bin/clang",
