+++ include/ethercat_interface/utilities.hpp
